refactor: 캐싱된 sharedPreference를 가져오는 로직을 변경
