--- app/src/main/java/woowacourse/movie/application/MovieApp.kt
@@ -1,15 +1,0 @@
-package woowacourse.movie.application
-
-import android.app.Application
-import woowacourse.movie.sharedpreference.SharedPreferences
-
-class MovieApp: Application() {
-    companion object {
-        lateinit var prefs: SharedPreferences
-    }
-
-    override fun onCreate() {
-        prefs = SharedPreferences(applicationContext)
-        super.onCreate()
-    }
-}
